Finished Bar Chart, started Pie Chart
Finished bar chart view. It is not generalized and will only plot average temps (I will go back and correct this if there is time). I have started the pie chart graph. I have planned it such that this pie chart graph can be used with any column from our data set. I have all of the data components in place for this class, now I just have to implement a method that draws the pieChart (aka the hard part).

+++ Thomas_Dickens_Proj8/DataImport.pde
@@ -48,9 +48,9 @@ class DataImport{
       int rMaxT = row.getInt("record_max_temp");
       int rMinTYr = row.getInt("record_min_temp_year");
       int rMaxTYr = row.getInt("record_max_temp_year");
-      float prec = row.getInt("actual_precipitation");
-      float aPrec = row.getInt("average_precipitation");
-      float rPrec = row.getInt("record_precipitation");
+      float prec = row.getFloat("actual_precipitation");
+      float aPrec = row.getFloat("average_precipitation");
+      float rPrec = row.getFloat("record_precipitation");
       
       dttm[i] = dtime;
       meanTemp[i] = meanT;
